requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic

langchain
langchain-core
langchain-openai
langchain-chroma
langgraph
langgraph-checkpoint-postgres

openai
google-generativeai
Pillow
langsmith

psycopg2-binary
psycopg2
psycopg
psycopg-pool
aiopg
tabulate
cryptography
pytest